# feat: enable fs.cachedChecks by default