### Description

`fs.cachedChecks` has been available and opt-in from vite@5.0.8.

See:
- https://github.com/vitejs/vite/pull/15279

In Windows, we have:
- tryFsResolve: 179.71ms -> 13.16ms (~14x faster)
- resolveId: 234ms -> 50ms (~4.7x faster)
- 5% overall

This PR switches the flag to be enabled by default.

We also consider that if there is custom watch ignore options, the user is opting into not watching these files not only HMR but also for resolving.

In future PRs we need to:
- move the public files fs caching implementation to use this new cache
- re-define the fs cache root to be the root of the monorepo instead of `root` to better support that use case
- decide if sharing cached trees is worth the extra complexity: https://github.com/vitejs/vite/pull/15294

---

### What is the purpose of this pull request? <!-- (put an "X" next to an item) -->

- [ ] Bug fix
- [ ] New Feature
- [ ] Documentation update
- [x] Other